fix: fix cyclic deps
