--- src/app/framework/i18n/models/language.ts
@@ -1,11 +1,0 @@
-export interface Language {
-  code: string;
-  name: string;
-  culture: string;
-}
-
-export const initialLanguage: Language = {
-  code: '',
-  name: '',
-  culture: ''
-};
